Move `Flume` to a new `.data._flumes` module
